apps/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-multipart==0.0.6
pydantic==2.8.2
python-dotenv==1.0.1
requests==2.32.3
msal==1.25.0
pymilvus==2.6.1
sentence-transformers==2.2.2
torch==2.1.1
pillow==10.4.0
azure-ai-documentintelligence==1.0.0b1
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-core==1.29.5
pytesseract==0.3.13
openai==1.44.1
tenacity==8.4.2
pywin32==306
